+++ src/main/java/io/github/wotjd243/ecommerce/user/application/ShippingAddressService.java
@@ -3,7 +3,7 @@ package io.github.wotjd243.ecommerce.user.application;
 import io.github.wotjd243.ecommerce.user.application.dto.ShippingAddressDto;
 import io.github.wotjd243.ecommerce.user.domain.ShippingAddress;
 import io.github.wotjd243.ecommerce.user.domain.ZipCode;
-import io.github.wotjd243.ecommerce.user.infra.ShippingAddressRepository;
+import io.github.wotjd243.ecommerce.user.domain.ShippingAddressRepository;
 
 public class ShippingAddressService {
 
@@ -13,8 +13,9 @@ public class ShippingAddressService {
         this.shippingAddressRepository = shippingAddressRepository;
     }
 
-    public ShippingAddress addShippingAddress(ShippingAddressDto shippingAddressDto) {
+    public ShippingAddress addShippingAddress(String userId, ShippingAddressDto shippingAddressDto) {
         ShippingAddress shippingAddress = new ShippingAddress(
+                userId,
                 shippingAddressDto.getAddress(),
                 ZipCode.codeOf(shippingAddressDto.getZipCode())
         );
